[update] 購物車金額資料型態調整，購物車邏輯調整

+++ src/EtheriT.Coker.Application.Shared/Dto/Order/OrderDetailsGetAllDto.cs
@@ -10,9 +10,9 @@ namespace EtheriT.Coker.Application.Shared.Dto.Order
         public string? ImagePath { get; set; }
         public string? S1Title { get; set; }
         public string? S2Title { get; set; }
-        public double Price { get; set; }
-        public int? SCPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal? SCPrice { get; set; }
         public int Quantity { get; set; }
-        public double Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
     }
 }
